libqcomtee: create private header
Move some shared symbols from qcomte_object.c.
Move tee_call_t refernce to the root object to keep life easier.
